requirements.txt
dnspython
twisted